src/lib/s2_recognition/s22_Neural_engine/cnn/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
numpy==1.26.4
opencv-python>=4.8
tqdm>=4.66
scikit-learn>=1.3
pyyaml>=6.0
onnx>=1.14
onnxruntime>=1.16; python_version < "3.13"
